requirements.txt
rest_framework
djangorestframework-simplejwt
djoser
django-cors-headers